pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "video-autocut"
version = "0.1.0"
description = "AI video analysis assistant powered by PydanticAI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "pydantic-ai>=0.0.70",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
video-autocut = "video_autocut.app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/video_autocut"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
